Fix issue with `CurrentDrop` with multiple drops
There is an issue with the `CurrentDrop` assignment when there is more than one near drop visible. `CurrentDrop` is always set to `nil` if there are multiple elements in the `DropsNear` table and the player is not close enough to the first element. 

To reproduce it, drop to items at close range, but not in the same drop. Now check to see if you can access both of them successfully.

The new version of the script will loop through all the `Drops`, then find all the `DropsNear` elements, and then set the `CurrentDrop` to the ID that is closest to the current player location. Tested it locally while creating drops from dead NPCs and it seems to work better. 

Feel free to change it as you see fit or comment on this PR if you have any suggestions.
Thanks!