Update main.lua
According to the Lua manual there are no shortcut operators in math.

This needs to be approved so I can put in the fix for the truck driving job inventory system.

The inventory is being saved to a json file so the shops' inventory can be maintained between server restarts.

This is done because the trucking job refills the inventory.

This bug was found and needs to be applied before the shops can be fixed.