Update mysql statements
If you use a MySQL server, you can use the "json" instead of "longtext". If you still use MariaDB, the "json" will be translated back to "longtext". So this only has an impact on the aesthetics. In the newest MySQL version, integer must not contain length anymore. Also, the "Primary Key" should always have an integer as value, which is why all keys are placed on the id, and e.g. "plate" is treated as a normal key.

At last I set the charset to UTF8MB4.