qb-inventory/server/main.lua
This simply checks the distance between players that are opening the inventories. It will reduce the amount of trolling caused by people that simply just open the inventory from afar. If the player is near them then they're going to see that person, they can easily who it is by looking at their id or something of that sorts. By that time it allows them to report them etc. Im still trying to figure out what other way you could check if they're doing it properly or not.