Update main.lua
This code was added to check the drop for items if no items this removes the bag drop from the world 
Adding more code on client side so your unable to pick up more than one drop at a time. Avoids getting a drop with or without items stuck to you by picking up another.